addresses issues in review
